chore: change output form
