+++ examples/package.json
@@ -17,7 +17,6 @@
     "@mui/material": "^5.4.2",
     "@realsee/dnalogel": "2.0.0-alpha.27-beta",
     "@realsee/five": "^5.0.0-alpha.127",
-    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.41",
     "@types/offscreencanvas": "^2019.6.4",
     "@types/react": "^17.0.33",
     "@types/react-dom": "^17.0.10",
@@ -29,7 +28,6 @@
     "eslint-config-alloy": "^4.5.1",
     "eslint-plugin-react": "^7.29.0",
     "hammerjs": "2.0.8",
-    "svelte": "^3.46.4",
     "three": "0.117.1",
     "typescript": "^4.5.5",
     "vite": "^2.8.0"
